Add test suite feature to run.py
run.py can once again run its full suite of tests. This is now its
default behaviour.

After every test, all the DB instances are terminated, the Q instance
is cleared, and new DB instances are started. This eliminates pending queue
entries from surviving one test to another, whether in the actual Q process
or in buffers in the DB processes.

This does not eliminate errors due to lingering state problems in the LB
instance. That is left for another time.

The commit also improves the wording of command-line help and fixes
a bug that prevented the --wait option from having any effect.