Fix errors in tests that respecify ndb or digest-length.
The code that modified the configurations for the LB and DB servers
was wrong. In fact, LB was never restarted.

This code cntralizes the changes to the configuration and restarts LB
in addition to DB (LB also needs to know the number of servers).
The new global dictionary 'active' contains the key parameters for
the currently-active test.